Changed the filters table's tweet_count column to be an integer instead of smallint, since we're already hitting the max value of smallint.

+++ static/sql/create_structural_dbs.sql
@@ -17,7 +17,7 @@ CREATE TABLE users
 CREATE TABLE filters(
   filter_name text PRIMARY KEY,
   last_tweet_timestamp timestamp DEFAULT CURRENT_TIMESTAMP,
-  tweet_count smallint
+  tweet_count integer
 );
 
 CREATE TABLE tweets(
